+++ src/Entities/AccessTokenEntityInterface.php
@@ -24,5 +24,5 @@ interface AccessTokenEntityInterface extends TokenInterface
     /**
      * Generate a string representation of the access token.
      */
-    public function __toString(): string;
+    public function toString(): string;
 }