+++ src/RequestEvent.php
@@ -12,10 +12,10 @@ declare(strict_types=1);
 
 namespace League\OAuth2\Server;
 
-use League\Event\Event;
+use League\OAuth2\Server\EventEmitting\AbstractEvent;
 use Psr\Http\Message\ServerRequestInterface;
 
-class RequestEvent extends Event
+class RequestEvent extends AbstractEvent
 {
     public const CLIENT_AUTHENTICATION_FAILED = 'client.authentication.failed';
     public const USER_AUTHENTICATION_FAILED = 'user.authentication.failed';