+++ src/ResponseTypes/BearerTokenResponse.php
@@ -31,7 +31,7 @@ class BearerTokenResponse extends AbstractResponseType
         $responseParams = [
             'token_type'   => 'Bearer',
             'expires_in'   => $expireDateTime - time(),
-            'access_token' => (string) $this->accessToken,
+            'access_token' => $this->accessToken->toString(),
         ];
 
         $refreshTokenPayload = json_encode([