+++ tests/Middleware/ResourceServerMiddlewareTest.php
@@ -33,12 +33,12 @@ class ResourceServerMiddlewareTest extends TestCase
 
         $accessToken = new AccessTokenEntity();
         $accessToken->setIdentifier('test');
-        $accessToken->setUserIdentifier(123);
+        $accessToken->setUserIdentifier('123');
         $accessToken->setExpiryDateTime((new DateTimeImmutable())->add(new DateInterval('PT1H')));
         $accessToken->setClient($client);
         $accessToken->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
-        $token = (string) $accessToken;
+        $token = $accessToken->toString();
 
         $request = (new ServerRequest())->withHeader('authorization', sprintf('Bearer %s', $token));
 
@@ -68,12 +68,12 @@ class ResourceServerMiddlewareTest extends TestCase
 
         $accessToken = new AccessTokenEntity();
         $accessToken->setIdentifier('test');
-        $accessToken->setUserIdentifier(123);
+        $accessToken->setUserIdentifier('123');
         $accessToken->setExpiryDateTime((new DateTimeImmutable())->sub(new DateInterval('PT1H')));
         $accessToken->setClient($client);
         $accessToken->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
-        $token = (string) $accessToken;
+        $token = $accessToken->toString();
 
         $request = (new ServerRequest())->withHeader('authorization', sprintf('Bearer %s', $token));